Refactored code to reduce line numbers
